+++ src/Symfony/Component/Mime/RawMessage.php
@@ -13,8 +13,6 @@ namespace Symfony\Component\Mime;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>
- *
- * @experimental in 4.3
  */
 class RawMessage implements \Serializable
 {
@@ -69,17 +67,11 @@ class RawMessage implements \Serializable
         $this->__unserialize(unserialize($serialized));
     }
 
-    /**
-     * @internal
-     */
     public function __serialize(): array
     {
         return [$this->message];
     }
 
-    /**
-     * @internal
-     */
     public function __unserialize(array $data): void
     {
         [$this->message] = $data;